+++ Modules/Domain/Tests/DomainTests/AuthUseCaseSpec.swift
@@ -1,6 +1,7 @@
-import Domain
 import DomainInterfaces
 import XCTest
+
+@testable import Domain
 
 final class AuthUseCaseSpec: XCTestCase {
 
@@ -58,30 +59,6 @@ final class AuthUseCaseSpec: XCTestCase {
 
         return (sut, spy)
     }
-
-}
-
-final class AuthUseCase: AuthUseCaseProtocol {
-    private let url: URL
-    private let httpClient: HttpPostClient
-
-    init(url: URL, httpClient: HttpPostClient) {
-        self.url = url
-        self.httpClient = httpClient
-    }
-
-    func auth(using user: User, completion: @escaping (Result<DomainInterfaces.Token, DomainError>) -> Void) {
-        httpClient.post(to: url, with: user.toData()) { result in
-            switch result {
-            case .success(let data):
-                guard let token: Token = data?.toModel() else { return completion(.failure(.genericError)) }
-                    completion(.success(token))
-            case .failure:
-                completion(.failure(.genericError))
-            }
-        }
-    }
-
 }
 
 final class HttpPostClientSpy: HttpPostClient {
